Move digital lane model and delegate in new package
