+++ caqtus/gui/condetrol/timelanes_editor/__init__.py
@@ -1,12 +1,12 @@
+from . import digital_lane_editor
 from ._delegate import TimeLaneDelegate
 from ._time_lanes_editor import TimeLanesEditor
 from .analog_lane_model import AnalogTimeLaneModel
-from .digital_lane_model import DigitalTimeLaneModel
 from .model import TimeLaneModel
 
 __all__ = [
     "TimeLanesEditor",
-    "DigitalTimeLaneModel",
+    "digital_lane_editor",
     "TimeLaneModel",
     "TimeLaneDelegate",
     "AnalogTimeLaneModel",
