+++ caqtus/gui/condetrol/timelanes_editor/digital_lane_editor/__init__.py
@@ -1,0 +1,4 @@
+from .digital_lane_delegate import DigitalTimeLaneDelegate
+from .digital_lane_model import DigitalTimeLaneModel
+
+__all__ = ["DigitalTimeLaneModel", "DigitalTimeLaneDelegate"]
